- working on 64bits port of jslibs - make data types conversions safer ( by using types like uint32_t instead of unsigned short ) - moduleManager.h: fix a 64bits port issue.

+++ moduleManager/moduleManager.h
@@ -12,7 +12,7 @@
  * License.
  * ***** END LICENSE BLOCK ***** */
 
-typedef unsigned int ModuleId;
+typedef unsigned long int ModuleId;
 
 typedef JSBool (*ModuleInitFunction)(JSContext *, JSObject *);
 typedef JSBool (*ModuleReleaseFunction)(JSContext *cx);
